requirements.txt
twelvelabs>=0.4.11
opencv-python==4.8.1.78
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0